flask-server/requirements.txt
Flask
Gunicorn
pillow
numpy
opencv-python
torch
matplotlib
Werkzeug
imgkit
timm
CORS
flask_cors